Extract suffix from Variable naming
Tensorflow adds ':0' to variable name but this makes it hard to find
existing variables as the true variable name is different from it.

So as variable names to be treated equaly in Theano and Tensorflow backend,
and `name` attribute to match with variabel name, we remove ':suffix'.

Optimizer slot namings are changed pattern so that variable names do
not corride when saving on HDF5 file.